Keep event normalization backend-neutral
