+++ src/mindroom/matrix/cache/sqlite_event_cache.py
@@ -14,6 +14,7 @@ import aiosqlite
 from mindroom.logging_config import get_logger
 
 from . import sqlite_event_cache_events, sqlite_event_cache_threads
+from .event_normalization import normalize_event_source_for_cache
 from .sqlite_agent_message_snapshot import load_sqlite_agent_message_snapshot
 
 if TYPE_CHECKING:
@@ -535,7 +536,7 @@ class SqliteEventCache:
         cached_at = time.time()
         events_by_room: dict[str, list[tuple[str, dict[str, Any]]]] = {}
         for event_id, room_id, event_data in events:
-            normalized_event = sqlite_event_cache_events.normalize_event_source_for_cache(event_data, event_id=event_id)
+            normalized_event = normalize_event_source_for_cache(event_data, event_id=event_id)
             events_by_room.setdefault(room_id, []).append((event_id, normalized_event))
 
         for room_id, room_events in events_by_room.items():
@@ -674,7 +675,7 @@ class SqliteEventCache:
 
     async def append_event(self, room_id: str, thread_id: str, event: dict[str, Any]) -> bool:
         """Append one event when the thread already has cached data."""
-        normalized_event = sqlite_event_cache_events.normalize_event_source_for_cache(event)
+        normalized_event = normalize_event_source_for_cache(event)
         return bool(
             await self._write_operation(
                 room_id,
