+++ tach.toml
@@ -956,7 +956,7 @@ path = "mindroom.matrix.cache"
 depends_on = [
     "mindroom.matrix.cache.agent_message_snapshot",
     "mindroom.matrix.cache.event_cache",
-    "mindroom.matrix.cache.sqlite_event_cache_events",
+    "mindroom.matrix.cache.event_normalization",
     "mindroom.matrix.cache.write_coordinator",
     "mindroom.matrix.cache.thread_cache_helpers",
     "mindroom.matrix.cache.thread_history_result",
@@ -980,11 +980,16 @@ visibility = ["mindroom.matrix.cache.sqlite_event_cache"]
 path = "mindroom.matrix.cache.sqlite_event_cache_threads"
 depends_on = [
     "mindroom.matrix.cache.event_cache",
+    "mindroom.matrix.cache.event_normalization",
     "mindroom.matrix.cache.sqlite_event_cache_events",
 ]
 
 [[modules]]
 path = "mindroom.matrix.cache.event_cache"
+depends_on = []
+
+[[modules]]
+path = "mindroom.matrix.cache.event_normalization"
 depends_on = []
 
 [[modules]]
@@ -1008,11 +1013,14 @@ depends_on = [ "mindroom.matrix.thread_bookkeeping"]
 
 [[modules]]
 path = "mindroom.matrix.cache.sqlite_event_cache_events"
-depends_on = ["mindroom.matrix.event_info"]
+depends_on = [
+    "mindroom.matrix.event_info",
+]
 
 [[modules]]
 path = "mindroom.matrix.cache.sqlite_event_cache"
 depends_on = [
+    "mindroom.matrix.cache.event_normalization",
     "mindroom.logging_config",
     "mindroom.matrix.cache.sqlite_agent_message_snapshot",
     "mindroom.matrix.cache.sqlite_event_cache_events",
@@ -1027,7 +1035,7 @@ depends_on = []
 [[modules]]
 path = "mindroom.matrix.cache.thread_writes"
 depends_on = [
-    "mindroom.matrix.cache.sqlite_event_cache_events",
+    "mindroom.matrix.cache.event_normalization",
     "mindroom.matrix.event_info",
     "mindroom.matrix.thread_bookkeeping",
 ]
